refactor publiclayout for better mantainability and readability
